+++ adminPanel/src/components/Comment.svelte
@@ -1,19 +1,8 @@
 <script>
+  import { formatDateTime } from "../utils";
+
   export let commentData;
   export let sendDeleteComment;
-
-  const formatDateTime = (inputDate) => {
-    const date = new Date(inputDate);
-
-    const day = String(date.getDate()).padStart(2, '0');
-    const month = String(date.getMonth() + 1).padStart(2, '0');
-    const year = String(date.getFullYear()).slice(-2);
-    const hours = String(date.getHours()).padStart(2, '0');
-    const minutes = String(date.getMinutes()).padStart(2, '0');
-    const seconds = String(date.getSeconds()).padStart(2, '0');
-
-    return `${day}.${month}.${year} ${hours}:${minutes}:${seconds}`;
-  }
 
   const deleteComment = async () => {
     const response = await fetch('http://' + import.meta.env.VITE_HOST + '/comment', {
@@ -22,7 +11,6 @@
       body: JSON.stringify({ comment_id: commentData.comment_id })
     });
     if (!response.ok) {
-      console.log(response);
       throw new Error('Network error');
     }
     sendDeleteComment(commentData.story_id);
